fresh: SoftwareUpdate: Switch to Fetch2 for update checking
Fully drop Lineage's checker (but keep the OSS licenses since the controller is kinda based on theirs) and use Fetch2 for manifest downloading as well.

+++ app/src/main/java/io/tenseventyseven/fresh/ota/db/CurrentSoftwareUpdate.java
@@ -56,7 +56,7 @@ public class CurrentSoftwareUpdate {
         editor.putInt(OTA_DOWNLOAD_STATE, UpdateDownload.OTA_DOWNLOAD_STATE_NOT_STARTED);
         editor.putBoolean(OTA_DOWNLOAD_VERIFIED, false);
         editor.putInt(OTA_DOWNLOAD_PROGRESS, 0);
-        editor.putInt(OTA_DOWNLOAD_ETA, 0);
+        editor.putLong(OTA_DOWNLOAD_ETA, 0);
         editor.commit();
     }
